pyproject.toml
[project]
name = "ai4stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "akshare>=1.18.26",
    "akshare-proxy-patch>=0.2.7",
    "backtrader>=1.9.78.123",
    "fastparquet>=2025.12.0",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandarallel>=1.6.5",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "tk>=0.1.0",
    "tqdm>=4.67.3",
    "tushare>=1.4.24",
    "yfinance>=1.2.0",
]
